test: gate permission-denied test behind //go:build !windows
Address Copilot review note on PR #2. The test was using a runtime
GOOS check to skip on Windows, which is fragile: it relies on the
test compiling on Windows in the first place (it does today because
os.Geteuid returns -1 there) and on mode 0o000 actually denying
reads (it doesn't on NTFS for the file owner).

Move the test to a Unix-only file with a build tag so the intent is
expressed structurally — the test simply doesn't exist for Windows
builds, and the runtime.GOOS import goes away.
